Reorganize things to minimize API surface
