+++ Sources/Dependencies/Dependencies.swift
@@ -42,7 +42,6 @@ public struct Dependencies {
     }
   }
 
-  private static var sharedAliases = Set<AnyHashable>()
   private var values = [ObjectIdentifier: DependencyValue]()
 
   public subscript<T>(_ key: T.Type) -> T.Value where T: DependencyKey {
@@ -64,40 +63,5 @@ public struct Dependencies {
   /// when using the library.
   public static func _new() -> Self {
     .init()
-  }
-
-  // MARK: Aliases
-
-  public func alias<Value>(_ keyPath: WritableKeyPath<Dependencies, Value>) -> DependencyAlias<Value>? {
-    for alias in Self.sharedAliases {
-      if let alias = alias as? DependencyAlias<Value>, alias.aliases.contains(keyPath) {
-        return alias
-      }
-    }
-    return nil
-  }
-
-  public mutating func synchronizeAliasedDependencies<Value>
-  (_ keyPath: WritableKeyPath<Dependencies, Value>) {
-    guard let alias = self.alias(keyPath) else { return }
-    let currentValue = self[keyPath: alias.default]
-    for keyPath in alias.aliases.filter({ $0 != alias.default }) {
-      self[keyPath: keyPath] = currentValue
-    }
-  }
-
-  public mutating func define<Value>(_ alias: DependencyAlias<Value>) {
-    remove(alias)
-    _ = Self.sharedAliases.insert(alias)
-  }
-
-  public mutating func remove<Value>(_ alias: DependencyAlias<Value>) {
-    if let existing = self.alias(alias.default) {
-      Self.sharedAliases.remove(existing)
-    }
-  }
-
-  public mutating func removeAllAliases() {
-    Self.sharedAliases.removeAll()
   }
 }
